requirements.txt
# Used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask>=1.0.0

chatterbot-corpus>=1.2.0
SQLAlchemy>=1.2
